+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/ide/completion/HuffOpcodeCompletionProvider.kt
@@ -110,6 +110,7 @@ class HuffOpcodeCompletionProvider : CompletionProvider<CompletionParameters>() 
             "not",
             "byte",
             "sha3",
+            "keccak256",
             "clz",
             // Environmental Information
             "address",